requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions 
requests 
xlrd
openpyxl==3.1.2 
pandas==2.1.2

